Training Security (#347)
## Overview
Training is a crucial feature for the Gonka network, but is not fully ready for usage. Development and experiments need to be possible, but general usage should be restricted, both to avoid DOS attacks and to avoid misunderstandings.

## Solution
All training related messages will be put behind two allow-lists for training. Only addresses in the list will be allowed to send training messages. The list will be controlled via governance.
The `EXEC` training list is for participants executing training.
The `START` allow list if for participants that can initiate and assign training.

## Implementation
1. The two new Training Allow list are added to the inference keeper as a `collections.KeySet`
2. Existing training messages are gated via the new method `CheckTrainingAllowList`. There are two lists of messages, one for each allow list.
3. New messages are added for removing an address, adding an address and setting the entire allow list. All are governance only messages and are verified to have the module Authority.
4. A query is added for checking what addresses are on the lists. Enums were not working with the CLI as expected so a pure number was used for the query to specify `EXEC` or `START`.

## Additional
There are some minor fixes in training code generally (for security audits) as well.