Unified permissions (#430)
* Add unified permissions
Instead of a variety of disparate approaches and policies, unify all permissions into permissions.go.
Each message is assigned a list of permissions required, and a method is made for verifying each permission.
ALL messages should immediately call CheckPermission. It will verify the message is authorized.
This is solely about permissions to send the message. This forces us to:
1. Think about the permission for every message (addresses several Certik issues and security risks)
2. Keep the list of permissions in one place

This does restrict some messages that were not previously restricted.

* Most fix tests
Except: InvalidateInference and RevalidateInference are not the usual pattern of permissions, they have to be sent by the EpochGroup policy account. For now we will keep the "old way", this just doesn't fit the pattern.

* Try and fix the Integration tests

* Lock versions in

* Fix build

* Add instructions for using permissions

* Fix a few issues found by Copilot

* Final changes:
- Yes, double up: Permissions both in `permissions.go` AND the call to CheckPermission. This gives visibility:
1. Where we need to see it to understand the message process.
2. A global overview of permissions for all messages (in permissions.go)

* Final changes:
- Yes, double up: Permissions both in `permissions.go` AND the call to CheckPermission. This gives visibility:
1. Where we need to see it to understand the message process.
2. A global overview of permissions for all messages (in permissions.go)

* A script for verifying coverage of all permissions