Switch to bytes for storage (#487)
This will be especially important when we switch to storing original payloads.
For now, it's an added layer of protection against encoding issues either in payloads given in the API or responses from LLMs, neither of which are completely controlled by the protocol, not generally predictable.
Storage now goes to bytes. For the entire data flow for responses and requests, we keep everything in bytes except for:
1. CanonicalizedJSON for modified prompts. Here, JSON serialization is needed, and produces strings.
2. Test input - for simplicity.

There were also some name changes for clarity.

This protects especially against malicious attempts to use non-standard encoding in payloads to try and break inference validations, a potentially disastrous impact on the chain.